code/requirements.txt
streamlit==1.31.1
openai==1.3.5
PyPDF2==3.0.1
python-dotenv==1.0.0
nltk==3.8.1
rank-bm25==0.2.2
numpy==1.24.3
